// src/modules/user/member/model/types.ts
export type GET_USER_MEMBER_REQUEST_TYPE = {
  startDate: string;
  endDate: string;
  idType: string;
  memberType: string;
  group: string;
  searchKeyword: string;
  searchType: string;
};

export type GET_MEMBER_IP_REQUEST_TYPE = {
  startDate: string;
  endDate: string;
  idType: string;
  memberType: string;
  searchKeyword: string;
  searchType: string;
};